+++ scalus-core/jvm/src/main/scala/scalus/uplc/builtin/JVMPlatformSpecific.scala
@@ -295,6 +295,11 @@ trait JVMPlatformSpecific extends PlatformSpecific {
         )
         ()
     }
+
+    override def createDirectories(path: String): Unit = {
+        Files.createDirectories(Paths.get(path))
+        ()
+    }
 }
 
 given PlatformSpecific = JVMPlatformSpecific